Move IP pinning logic to a separate file
