+++ src/runtime/server/middleware/session/index.ts
@@ -1,9 +1,9 @@
-import { H3Event, eventHandler, setCookie, parseCookies, deleteCookie } from 'h3'
+import { deleteCookie, eventHandler, H3Event, parseCookies, setCookie } from 'h3'
 import { nanoid } from 'nanoid'
 import dayjs from 'dayjs'
-import { hash, argon2id, verify } from 'argon2'
 import type { SameSiteOptions } from '../../../../module'
 import { dropStorageSession, getStorageSession, setStorageSession } from './storage'
+import { getRequestIpAddress, hashIpAddress, ipAddressesMatch } from './ipPinning'
 import { useRuntimeConfig } from '#imports'
 
 const SESSION_COOKIE_NAME = 'sessionId'
@@ -23,58 +23,6 @@ export declare interface Session {
   createdAt: Date
   ip?: string
   [key: string]: any
-}
-
-const argon2Options = {
-  // cryptographically-secure salt is generated automatically
-  type: argon2id, // resistant against GPU & tradeoff attacks
-  hashLength: 60
-}
-
-/**
- * Get a hashed representation of the given IP address (for GDPR compliance)
- * @param ip string|undefined The IP address to hash
- */
-const hashIpAddress = (ip: string|undefined): Promise<string|undefined> =>
-  !ip
-    ? Promise.resolve(undefined)
-    : hash(ip, argon2Options)
-
-/**
- * Check that the given (raw) IP address and the hashed IP address match
- * @param ip string|undefined The IP address to verify
- * @param ipHash string|undefined The (hashed) IP address to test against
- */
-const ipAddressesMatch = (ip: string|undefined, ipHash: string|undefined): Promise<boolean> => !(ip && ipHash) ? Promise.resolve(false) : verify(ipHash, ip, argon2Options)
-
-/**
- * Extract the IP address from an HTTP header
- * @param header string|string[]|undefined The header value to inspect and extract the IP from
- */
-const extractIpFromHeader = (header?: string|string[]): string|undefined => {
-  if (Array.isArray(header)) {
-    return header[0].split(',')[0]
-  }
-
-  if (typeof header === 'string') {
-    return header.split(',')[0]
-  }
-
-  return undefined
-}
-
-/**
- * Get the IP address corresponding to the user's request
- * @param event H3Event Event passing through middleware
- */
-const getRequestIpAddress = ({ req }: H3Event): string|undefined => {
-  const foundIp = [
-    'x-forwarded-for',
-    'true-client-ip',
-    'cf-connecting-ip'
-  ].find(headerName => extractIpFromHeader(req.headers[headerName]))
-
-  return foundIp ?? req.connection?.remoteAddress ?? req.socket.remoteAddress
 }
 
 /**
